Rename code_analyzer to code_hasher
