+++ bionic/code_hasher.py
